+++ client/src/javascript/intl-phone-number-field-default-from-browser.js
@@ -1,0 +1,1 @@
+import "./components/load-default-country-from-browser";